FIX : INTEGER TO LONG

+++ backend/src/main/java/com/ygss/backend/user/dto/EditUserInfoResponseDto.java
@@ -10,6 +10,6 @@ public class EditUserInfoResponseDto {
     private String email;
     private String password;
     private Integer workedAt;
-    private Integer salary;
-    private Integer totalRetirePension;
+    private Long salary;
+    private Long totalRetirePension;
 }
